refactor: Session, JWT 환경 분리
